lib: put cufinufft wrapper function declarations in their own header file
