+++ lib/cufinufft_wrapper.cc
@@ -1,29 +1,12 @@
-#ifndef _JAX_FINUFFT_GPU_H_
-#define _JAX_FINUFFT_GPU_H_
-
 #include <complex>
 
 #include "cufinufft.h"
 
+#include "cufinufft_wrapper.h"
+
 namespace jax_finufft {
 
 namespace gpu {
-
-template <typename T>
-struct plan_type;
-
-template <>
-struct plan_type<double> {
-  typedef cufinufft_plan type;
-};
-
-template <>
-struct plan_type<float> {
-  typedef cufinufftf_plan type;
-};
-
-template <typename T>
-void default_opts(cufinufft_opts* opts);
 
 template <>
 void default_opts<float>(cufinufft_opts* opts) {
@@ -34,9 +17,6 @@ template <>
 void default_opts<double>(cufinufft_opts* opts) {
   cufinufft_default_opts(opts);
 }
-
-template <typename T>
-void update_opts(cufinufft_opts* opts, int dim, cudaStream_t stream);
 
 template <>
 void update_opts<float>(cufinufft_opts* opts, int dim, cudaStream_t stream) {
@@ -55,10 +35,6 @@ void update_opts<double>(cufinufft_opts* opts, int dim, cudaStream_t stream) {
   }
 }
 
-template <typename T>
-int makeplan(int type, int dim, const int64_t nmodes[3], int iflag, int ntr, T eps,
-             typename plan_type<T>::type* plan, cufinufft_opts* opts);
-
 template <>
 int makeplan<float>(int type, int dim, const int64_t nmodes[3], int iflag, int ntr, float eps,
                     typename plan_type<float>::type* plan, cufinufft_opts* opts) {
@@ -74,10 +50,6 @@ int makeplan<double>(int type, int dim, const int64_t nmodes[3], int iflag, int 
   return cufinufft_makeplan(type, dim, tmp_nmodes, iflag, ntr, eps, plan, opts);
 }
 
-template <typename T>
-int setpts(typename plan_type<T>::type plan, int64_t M, T* x, T* y, T* z, int64_t N, T* s, T* t,
-           T* u);
-
 template <>
 int setpts<float>(typename plan_type<float>::type plan, int64_t M, float* x, float* y, float* z,
                   int64_t N, float* s, float* t, float* u) {
@@ -89,9 +61,6 @@ int setpts<double>(typename plan_type<double>::type plan, int64_t M, double* x, 
                    double* z, int64_t N, double* s, double* t, double* u) {
   return cufinufft_setpts(plan, M, x, y, z, N, s, t, u);
 }
-
-template <typename T>
-int execute(typename plan_type<T>::type plan, std::complex<T>* c, std::complex<T>* f);
 
 template <>
 int execute<float>(typename plan_type<float>::type plan, std::complex<float>* c,
@@ -109,9 +78,6 @@ int execute<double>(typename plan_type<double>::type plan, std::complex<double>*
   return cufinufft_execute(plan, _c, _f);
 }
 
-template <typename T>
-void destroy(typename plan_type<T>::type plan);
-
 template <>
 void destroy<float>(typename plan_type<float>::type plan) {
   cufinufftf_destroy(plan);
@@ -122,11 +88,6 @@ void destroy<double>(typename plan_type<double>::type plan) {
   cufinufft_destroy(plan);
 }
 
-template <int ndim, typename T>
-T* y_index(T* y, int64_t index) {
-  return &(y[index]);
-}
-
 template <>
 double* y_index<1, double>(double* y, int64_t index) {
   return NULL;
@@ -134,11 +95,6 @@ double* y_index<1, double>(double* y, int64_t index) {
 
 template <>
 float* y_index<1, float>(float* y, int64_t index) {
-  return NULL;
-}
-
-template <int ndim, typename T>
-T* z_index(T* z, int64_t index) {
   return NULL;
 }
 
@@ -155,5 +111,3 @@ float* z_index<3, float>(float* z, int64_t index) {
 }  // namespace gpu
 
 }  // namespace jax_finufft
-
-#endif
