+++ lib/jax_finufft_gpu.cc
@@ -2,7 +2,7 @@
 // It is exposed as a standard pybind11 module defining "capsule" objects containing our
 // method. For simplicity, we export a separate capsule for each supported dtype.
 
-#include "cufinufft.h"
+#include "cufinufft_wrapper.h"
 #include "kernels.h"
 #include "pybind11_kernel_helpers.h"
